Add save system and update level assets
Introduced a new save system with AllGameData, PlayerData, and JsonSaveLoad scripts. Renamed and added level assets to support multiple levels. Updated font asset to include additional glyphs and kerning for numbers. Modified several manager and UI scripts to integrate with the new save and level systems. BUG Found: CaveMan cannot hit enemy ninja

+++ Assets/Scripts/ScriptableObjects/LevelDataSO.cs
@@ -24,6 +24,6 @@ namespace ScriptableObjects
     {
         public FighterType FighterType;
         public int UnitSpawnCount;
-        public int TimeBetweenEachUnits;
+        public float TimeBetweenEachUnits;
     }
 }
